[FE] feat: 구매상품 뷰 추가 #115
